src/TeamsOrchestrator/requirements.txt
# Azure AI Agent Server for hosted agents
azure-ai-agentserver-agentframework==1.0.0b3
agent-framework

# Azure AI integration
azure-ai-projects==1.0.0b6
azure-ai-inference==1.0.0b5
azure-identity

# Additional utilities
python-dotenv
pydantic
pydantic-settings

# Original dependencies  
openai

# Agent SDK dependencies
aiohttp
microsoft-agents-hosting-aiohttp
microsoft-agents-hosting-core
microsoft-agents-authentication-msal
microsoft-agents-activity

# Note: Azure Developer CLI (azd) is not available via pip
# Run setup_deployment.sh or install manually with:
# curl -fsSL https://aka.ms/install-azd.sh | bash